Add layout with menu and restructure context
